refactor. 바뀐 Exception 정책에 따라 기존 Exception 수정 및 엔티티 클래스들 한번에 모아놓도록 패키지 구조 수정
