+++ BE/src/main/java/team07/airbnb/common/auth/jwt/JwtUserDetails.java
@@ -7,7 +7,6 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import team07.airbnb.domain.user.dto.TokenUserInfo;
-import team07.airbnb.domain.user.entity.UserEntity;
 
 import java.util.Collection;
 import java.util.Collections;
