+++ BE/src/main/java/team07/airbnb/domain/accommodation/AccommodationService.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.locationtech.jts.geom.Point;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import team07.airbnb.domain.accommodation.entity.AccommodationEntity;
+import team07.airbnb.entity.AccommodationEntity;
 import team07.airbnb.common.exception.UnAuthorizedException;
-import team07.airbnb.domain.user.entity.UserEntity;
+import team07.airbnb.entity.UserEntity;
 import team07.airbnb.domain.user.enums.Role;
 import team07.airbnb.domain.user.service.UserService;
 import team07.airbnb.common.util.GeometryHelper;
