+++ BE/src/main/java/team07/airbnb/domain/booking/BookingRepository.java
@@ -2,8 +2,8 @@ package team07.airbnb.domain.booking;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import team07.airbnb.domain.booking.entity.BookingEntity;
-import team07.airbnb.domain.user.entity.UserEntity;
+import team07.airbnb.entity.BookingEntity;
+import team07.airbnb.entity.UserEntity;
 
 import java.util.List;
