+++ BE/src/main/java/team07/airbnb/domain/booking/dto/request/BookingRequest.java
@@ -3,9 +3,6 @@ package team07.airbnb.domain.booking.dto.request;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import org.springframework.lang.Nullable;
-import team07.airbnb.domain.booking.entity.BookingEntity;
-import team07.airbnb.domain.booking.property.BookingStatus;
-import team07.airbnb.domain.user.entity.UserEntity;
 
 import java.time.LocalDate;
