+++ BE/src/main/java/team07/airbnb/domain/discount/DiscountPolicyService.java
@@ -1,13 +1,11 @@
 package team07.airbnb.domain.discount;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.stereotype.Service;
 import team07.airbnb.domain.discount.beans.DiscountPolicy;
-import team07.airbnb.domain.discount.entity.DiscountPolicyEntity;
-import team07.airbnb.domain.discount.exception.UnknownDiscountPolicyException;
+import team07.airbnb.entity.DiscountPolicyEntity;
+import team07.airbnb.domain.discount.exception.DiscountPolicyNotFoundException;
 
 import java.util.Optional;
 
@@ -21,7 +19,7 @@ public class DiscountPolicyService {
     public long getDiscountPrice(long roughPrice) {
         //일단 주단위 할인
         Optional<DiscountPolicyEntity> weekDiscount = discountPolicyRepository.findByDescription("주단위할인");
-        DiscountPolicyEntity entity = weekDiscount.orElseThrow(UnknownDiscountPolicyException::new);
+        DiscountPolicyEntity entity = weekDiscount.orElseThrow(DiscountPolicyNotFoundException::new);
 
         String policyBeanName = entity.getPolicyBeanName();
         DiscountPolicy bean = (DiscountPolicy) ac.getBean(policyBeanName);
