--- BE/src/main/java/team07/airbnb/domain/discount/exception/UnknownDiscountPolicyException.java
@@ -1,7 +1,0 @@
-package team07.airbnb.domain.discount.exception;
-
-public class UnknownDiscountPolicyException extends RuntimeException {
-    public UnknownDiscountPolicyException() {
-        super("해당하는 할인 정책이 존재하지 않습니다");
-    }
-}
