+++ BE/src/main/java/team07/airbnb/domain/payment/PaymentRepository.java
@@ -2,6 +2,7 @@ package team07.airbnb.domain.payment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import team07.airbnb.entity.PaymentEntity;
 
 @Repository
 public interface PaymentRepository extends JpaRepository<PaymentEntity, Long> {
