+++ BE/src/main/java/team07/airbnb/domain/payment/PaymentService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import team07.airbnb.domain.booking.dto.BookingInfo;
+import team07.airbnb.entity.PaymentEntity;
 
 @Service
 @RequiredArgsConstructor
