+++ BE/src/main/java/team07/airbnb/domain/product/ProductStatus.java
@@ -1,4 +1,4 @@
-package team07.airbnb.domain.product.entity;
+package team07.airbnb.domain.product;
 
 public enum ProductStatus {
     OPEN, BOOKED, CLOSE
