+++ BE/src/main/java/team07/airbnb/domain/user/dto/TokenUserInfo.java
@@ -1,6 +1,6 @@
 package team07.airbnb.domain.user.dto;
 
-import team07.airbnb.domain.user.entity.UserEntity;
+import team07.airbnb.entity.UserEntity;
 import team07.airbnb.domain.user.enums.Role;
 
 public record TokenUserInfo(Long id, String name, String profileImg, Role role) {
