+++ BE/src/main/java/team07/airbnb/domain/user/repository/UserRepository.java
@@ -2,7 +2,7 @@ package team07.airbnb.domain.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import team07.airbnb.domain.user.entity.UserEntity;
+import team07.airbnb.entity.UserEntity;
 
 import java.util.Optional;
