+++ BE/src/main/java/team07/airbnb/entity/AccommodationEntity.java
@@ -1,4 +1,4 @@
-package team07.airbnb.domain.accommodation.entity;
+package team07.airbnb.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.CascadeType;
@@ -17,14 +17,9 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import team07.airbnb.domain.BaseEntity;
 import team07.airbnb.domain.accommodation.property.AccommodationLocation;
 import team07.airbnb.domain.accommodation.property.AccommodationType;
 import team07.airbnb.domain.accommodation.property.RoomInformation;
-import team07.airbnb.domain.booking.entity.BookingEntity;
-import team07.airbnb.domain.product.entity.ProductEntity;
-import team07.airbnb.domain.review.ReviewEntity;
-import team07.airbnb.domain.user.entity.UserEntity;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
