+++ BE/src/main/java/team07/airbnb/entity/BookingEntity.java
@@ -1,11 +1,10 @@
-package team07.airbnb.domain.booking.entity;
+package team07.airbnb.entity;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.OneToOne;
@@ -15,12 +14,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import team07.airbnb.domain.BaseEntity;
 import team07.airbnb.domain.booking.property.BookingStatus;
-import team07.airbnb.domain.payment.PaymentEntity;
-import team07.airbnb.domain.product.entity.ProductEntity;
-import team07.airbnb.domain.review.ReviewEntity;
-import team07.airbnb.domain.user.entity.UserEntity;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
