+++ BE/src/main/java/team07/airbnb/entity/DiscountPolicyEntity.java
@@ -1,4 +1,4 @@
-package team07.airbnb.domain.discount.entity;
+package team07.airbnb.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -7,7 +7,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
-import team07.airbnb.domain.BaseEntity;
+import team07.airbnb.entity.BaseEntity;
 
 @Entity
 @Table(name = "DISCOUNT_POLICY")
