+++ BE/src/main/java/team07/airbnb/entity/Pictures.java
@@ -1,4 +1,4 @@
-package team07.airbnb.domain.accommodation.entity;
+package team07.airbnb.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
