+++ BE/src/main/java/team07/airbnb/entity/ReplyEntity.java
@@ -1,4 +1,4 @@
-package team07.airbnb.domain.review;
+package team07.airbnb.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -6,7 +6,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import team07.airbnb.domain.user.entity.UserEntity;
 
 @Entity
 @Table(name = "REVIEW_REPLY")
